refactor: migrate RecruitmentItem folder to typescript
(cherry picked from commit 8fff3d3a5d11079c163a8ad5711909f109ab591a)
